masfro-backend-v2/pyproject.toml
[project]
name = "masfro-backend-v2"
version = "2.0.0"
description = "Multi-Agent System for Flood Route Optimization Backend v2 with Qwen 3 LLM Integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.0",
    "matplotlib>=3.10.6",
    "networkx>=3.4.2",
    "osmnx>=2.0.6",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
    "selenium>=4.36.0",
    "uvicorn>=0.37.0",
    "webdriver-manager>=4.0.2",
    "python-dotenv>=1.0.1",
    "bs4>=0.0.2",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.44",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.17.1",
    "rasterio>=1.4.3",
    "numpy>=2.2.6",
    "pyproj>=3.7.1",
    "lxml>=6.0.2",
    "rich>=14.2.0",
    "spacy>=3.8.11",
    "joblib>=1.5.2",
    "tenacity>=8.2.3",
    "pyyaml>=6.0.0",
    # v2: Qwen 3 LLM Integration via Ollama
    "ollama>=0.4.0",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
